repl: move markdown, image, and table to their own files
