+++ crates/repl/src/outputs.rs
@@ -1,27 +1,22 @@
 use std::time::Duration;
 
 use crate::stdio::TerminalOutput;
-use anyhow::Result;
 
 use gpui::{
-    percentage, Animation, AnimationExt, AnyElement, ClipboardItem, FontWeight, Render, Task,
-    TextRun, Transformation, View,
+    percentage, Animation, AnimationExt, AnyElement, ClipboardItem, FontWeight, Render,
+    Transformation, View,
 };
-use runtimelib::datatable::TableSchema;
-use runtimelib::media::datatable::TabularDataResource;
 use runtimelib::{ExecutionState, JupyterMessageContent, MimeBundle, MimeType};
-use serde_json::Value;
-use settings::Settings;
-use theme::ThemeSettings;
 use ui::{div, prelude::*, v_flex, IntoElement, Styled, Tooltip, ViewContext};
 
 mod image;
 use image::ImageView;
 
-use markdown_preview::{
-    markdown_elements::ParsedMarkdown, markdown_parser::parse_markdown,
-    markdown_renderer::render_markdown_block,
-};
+mod markdown;
+use markdown::MarkdownView;
+
+mod table;
+use table::TableView;
 
 /// When deciding what to render from a collection of mediatypes, we need to rank them in order of importance
 fn rank_mime_type(mimetype: &MimeType) -> usize {
@@ -39,234 +34,6 @@ fn rank_mime_type(mimetype: &MimeType) -> usize {
 pub(crate) trait SupportsClipboard {
     fn clipboard_content(&self, cx: &WindowContext) -> Option<ClipboardItem>;
     fn has_clipboard_content(&self, cx: &WindowContext) -> bool;
-}
-
-/// TableView renders a static table inline in a buffer.
-/// It uses the https://specs.frictionlessdata.io/tabular-data-resource/ specification for data interchange.
-pub struct TableView {
-    pub table: TabularDataResource,
-    pub widths: Vec<Pixels>,
-    cached_clipboard_content: ClipboardItem,
-}
-
-fn cell_content(row: &Value, field: &str) -> String {
-    match row.get(&field) {
-        Some(Value::String(s)) => s.clone(),
-        Some(Value::Number(n)) => n.to_string(),
-        Some(Value::Bool(b)) => b.to_string(),
-        Some(Value::Array(arr)) => format!("{:?}", arr),
-        Some(Value::Object(obj)) => format!("{:?}", obj),
-        Some(Value::Null) | None => String::new(),
-    }
-}
-
-// Declare constant for the padding multiple on the line height
-const TABLE_Y_PADDING_MULTIPLE: f32 = 0.5;
-
-impl TableView {
-    pub fn new(table: TabularDataResource, cx: &mut WindowContext) -> Self {
-        let mut widths = Vec::with_capacity(table.schema.fields.len());
-
-        let text_system = cx.text_system();
-        let text_style = cx.text_style();
-        let text_font = ThemeSettings::get_global(cx).buffer_font.clone();
-        let font_size = ThemeSettings::get_global(cx).buffer_font_size;
-        let mut runs = [TextRun {
-            len: 0,
-            font: text_font,
-            color: text_style.color,
-            background_color: None,
-            underline: None,
-            strikethrough: None,
-        }];
-
-        for field in table.schema.fields.iter() {
-            runs[0].len = field.name.len();
-            let mut width = text_system
-                .layout_line(&field.name, font_size, &runs)
-                .map(|layout| layout.width)
-                .unwrap_or(px(0.));
-
-            let Some(data) = table.data.as_ref() else {
-                widths.push(width);
-                continue;
-            };
-
-            for row in data {
-                let content = cell_content(&row, &field.name);
-                runs[0].len = content.len();
-                let cell_width = cx
-                    .text_system()
-                    .layout_line(&content, font_size, &runs)
-                    .map(|layout| layout.width)
-                    .unwrap_or(px(0.));
-
-                width = width.max(cell_width)
-            }
-
-            widths.push(width)
-        }
-
-        let cached_clipboard_content = Self::create_clipboard_content(&table);
-
-        Self {
-            table,
-            widths,
-            cached_clipboard_content: ClipboardItem::new_string(cached_clipboard_content),
-        }
-    }
-
-    fn escape_markdown(s: &str) -> String {
-        s.replace('|', "\\|")
-            .replace('*', "\\*")
-            .replace('_', "\\_")
-            .replace('`', "\\`")
-            .replace('[', "\\[")
-            .replace(']', "\\]")
-            .replace('<', "&lt;")
-            .replace('>', "&gt;")
-    }
-
-    fn create_clipboard_content(table: &TabularDataResource) -> String {
-        let data = match table.data.as_ref() {
-            Some(data) => data,
-            None => &Vec::new(),
-        };
-        let schema = table.schema.clone();
-
-        let mut markdown = format!(
-            "| {} |\n",
-            table
-                .schema
-                .fields
-                .iter()
-                .map(|field| field.name.clone())
-                .collect::<Vec<_>>()
-                .join(" | ")
-        );
-
-        markdown.push_str("|---");
-        for _ in 1..table.schema.fields.len() {
-            markdown.push_str("|---");
-        }
-        markdown.push_str("|\n");
-
-        let body = data
-            .iter()
-            .map(|record: &Value| {
-                let row_content = schema
-                    .fields
-                    .iter()
-                    .map(|field| Self::escape_markdown(&cell_content(record, &field.name)))
-                    .collect::<Vec<_>>();
-
-                row_content.join(" | ")
-            })
-            .collect::<Vec<String>>();
-
-        for row in body {
-            markdown.push_str(&format!("| {} |\n", row));
-        }
-
-        markdown
-    }
-
-    pub fn render(&self, cx: &ViewContext<ExecutionView>) -> AnyElement {
-        let data = match &self.table.data {
-            Some(data) => data,
-            None => return div().into_any_element(),
-        };
-
-        let mut headings = serde_json::Map::new();
-        for field in &self.table.schema.fields {
-            headings.insert(field.name.clone(), Value::String(field.name.clone()));
-        }
-        let header = self.render_row(&self.table.schema, true, &Value::Object(headings), cx);
-
-        let body = data
-            .iter()
-            .map(|row| self.render_row(&self.table.schema, false, &row, cx));
-
-        v_flex()
-            .id("table")
-            .overflow_x_scroll()
-            .w_full()
-            .child(header)
-            .children(body)
-            .into_any_element()
-    }
-
-    pub fn render_row(
-        &self,
-        schema: &TableSchema,
-        is_header: bool,
-        row: &Value,
-        cx: &ViewContext<ExecutionView>,
-    ) -> AnyElement {
-        let theme = cx.theme();
-
-        let line_height = cx.line_height();
-
-        let row_cells = schema
-            .fields
-            .iter()
-            .zip(self.widths.iter())
-            .map(|(field, width)| {
-                let container = match field.field_type {
-                    runtimelib::datatable::FieldType::String => div(),
-
-                    runtimelib::datatable::FieldType::Number
-                    | runtimelib::datatable::FieldType::Integer
-                    | runtimelib::datatable::FieldType::Date
-                    | runtimelib::datatable::FieldType::Time
-                    | runtimelib::datatable::FieldType::Datetime
-                    | runtimelib::datatable::FieldType::Year
-                    | runtimelib::datatable::FieldType::Duration
-                    | runtimelib::datatable::FieldType::Yearmonth => v_flex().items_end(),
-
-                    _ => div(),
-                };
-
-                let value = cell_content(row, &field.name);
-
-                let mut cell = container
-                    .min_w(*width + px(22.))
-                    .w(*width + px(22.))
-                    .child(value)
-                    .px_2()
-                    .py((TABLE_Y_PADDING_MULTIPLE / 2.0) * line_height)
-                    .border_color(theme.colors().border);
-
-                if is_header {
-                    cell = cell.border_1().bg(theme.colors().border_focused)
-                } else {
-                    cell = cell.border_1()
-                }
-                cell
-            })
-            .collect::<Vec<_>>();
-
-        let mut total_width = px(0.);
-        for width in self.widths.iter() {
-            // Width fudge factor: border + 2 (heading), padding
-            total_width += *width + px(22.);
-        }
-
-        h_flex()
-            .w(total_width)
-            .children(row_cells)
-            .into_any_element()
-    }
-}
-
-impl SupportsClipboard for TableView {
-    fn clipboard_content(&self, _cx: &WindowContext) -> Option<ClipboardItem> {
-        Some(self.cached_clipboard_content.clone())
-    }
-
-    fn has_clipboard_content(&self, _cx: &WindowContext) -> bool {
-        true
-    }
 }
 
 /// Userspace error from the kernel
@@ -311,72 +78,6 @@ impl ErrorView {
                 )
                 .into_any_element(),
         )
-    }
-}
-
-pub struct MarkdownView {
-    raw_text: String,
-    contents: Option<ParsedMarkdown>,
-    parsing_markdown_task: Option<Task<Result<()>>>,
-}
-
-impl MarkdownView {
-    pub fn from(text: String, cx: &mut ViewContext<Self>) -> Self {
-        let task = cx.spawn(|markdown_view, mut cx| {
-            let text = text.clone();
-            let parsed = cx
-                .background_executor()
-                .spawn(async move { parse_markdown(&text, None, None).await });
-
-            async move {
-                let content = parsed.await;
-
-                markdown_view.update(&mut cx, |markdown, cx| {
-                    markdown.parsing_markdown_task.take();
-                    markdown.contents = Some(content);
-                    cx.notify();
-                })
-            }
-        });
-
-        Self {
-            raw_text: text.clone(),
-            contents: None,
-            parsing_markdown_task: Some(task),
-        }
-    }
-}
-
-impl SupportsClipboard for MarkdownView {
-    fn clipboard_content(&self, _cx: &WindowContext) -> Option<ClipboardItem> {
-        Some(ClipboardItem::new_string(self.raw_text.clone()))
-    }
-
-    fn has_clipboard_content(&self, _cx: &WindowContext) -> bool {
-        true
-    }
-}
-
-impl Render for MarkdownView {
-    fn render(&mut self, cx: &mut ViewContext<Self>) -> impl IntoElement {
-        let Some(parsed) = self.contents.as_ref() else {
-            return div().into_any_element();
-        };
-
-        let mut markdown_render_context =
-            markdown_preview::markdown_renderer::RenderContext::new(None, cx);
-
-        v_flex()
-            .gap_3()
-            .py_4()
-            .children(parsed.children.iter().map(|child| {
-                div().relative().child(
-                    div()
-                        .relative()
-                        .child(render_markdown_block(child, &mut markdown_render_context)),
-                )
-            }))
-            .into_any_element()
     }
 }
